DOPS-1823: feat: add js build
