--- packages/js/interface/Slider/style.scss
@@ -1,60 +1,0 @@
-@import '../../../scss/setup.scss';
-
-//
-// JavaScript functional styles.
-//
-
-[data-slider] {
-	position: relative;
-	overflow-x: scroll;
-	-webkit-overflow-scrolling: touch;
-
-	@include breakpoint( tablet-min ) {
-		@include supports-grid {
-			overflow-x: hidden;
-		}
-	}
-
-}
-
-[data-slider-nav] {
-	cursor: pointer;
-	position: absolute;
-	height: 100%;
-	z-index: 1000;
-	display: none;
-
-	@include breakpoint( tablet-min ) {
-		@include supports-grid {
-			display: block;
-		}
-	}
-
-	// Nav should be visible on tablets.
-	@include breakpoint( desktop ) {
-		opacity: 0;
-
-		[data-slider]:hover & {
-			opacity: 1;
-		}
-	}
-}
-
-[data-slider-track] {
-
-	transition: $transition-transform;
-	will-change: transform;
-
-	// Fallback Flex layout.
-	// The component instance will likely override this with CSS Grid.
-	display: flex;
-
-	> * {
-		width: 100%;
-		min-width: calc( 1vw + #{rem-calc( 260 )} );
-
-		// Note: This is necessary to correctly calculate the item width in JS, for now.
-		// Additional notes within getOffset in Slider.js;
-		padding-right: calc( 1vw + 1rem );
-	}
-}
