--- packages/js/interface/TabsManager/index.js
@@ -1,9 +1,0 @@
-import './style.scss';
-import TabsManager from './TabsManager.js';
-
-( function( $ ) {
-
-	const tabManager = new TabsManager();
-	tabManager.init();
-
-}( jQuery ) );
