+++ packages/js/src/common.js
@@ -1,0 +1,11 @@
+import initDropdowns from './interface/Dropdown/index';
+
+const onSafeResize = function() {
+	const width = window.innerWidth;
+
+	initDropdowns( width );
+};
+
+window.addEventListener( 'load', () => {
+	onSafeResize();
+});
