+++ packages/js/src/interface/Dropdown/index.js
@@ -3,12 +3,14 @@ import Dropdown from './Dropdown.js';
 // Initialize all Dropdowns.
 
 export default function() {
-	const dropdowns = [ ... document.querySelectorAll( '.js-Dropdown' ) ];
+	const dropdowns = [ ... document.querySelectorAll( '.pmc-js-Dropdown' ) ];
 	const width     = window.innerWidth;
-
+	
 	dropdowns.forEach( el => {
 		if ( 768 > width && undefined === el.pmcDropdown ) {
 			new Dropdown( el );
+			console.log( el );
+			
 		} else if ( 768 <= width && undefined !== el.pmcDropdown ) {
 			el.pmcDropdown.destroy();
 		}
