+++ packages/js/src/utils/ajaxIconSprite.js
