+++ packages/js/webpack.config.js
@@ -1,0 +1,11 @@
+const path = require('path');
+
+module.exports = {
+	entry: './src/common.js',
+	watch: true,
+	mode: 'development',
+	output: {
+		filename: 'common.js',
+		path: path.resolve(__dirname, 'build')
+	}
+};
