+++ packages/larva/lib/server.js
@@ -26,6 +26,7 @@ app.use( express.static( 'build' ) );
 // TODO: these will be updated to paths that point to a node module for use out of the mono-repo
 app.use( '/utils' , express.static( path.join( appConfiguration.larvaPatternsDir, '../css-utilities/build/css' ) ) );
 app.use( '/icons' , express.static( path.join( appConfiguration.larvaPatternsDir, '../icons/build' ) ) );
+app.use( '/js' , express.static( path.join( appConfiguration.larvaPatternsDir, '../js/build' ) ) );
 app.use( '/algos' , express.static( path.join( appConfiguration.larvaPatternsDir, '../css-algorithms/build/css' ) ) );
 app.use( '/patterns' , express.static( appConfiguration.larvaPatternsDir ) ); // should point to consuming project dir
 app.use( '/static' , express.static( path.join( __dirname, '../static' ) ) );
