+++ packages/patterns/pattern.html
@@ -11,6 +11,8 @@
 		<link rel="stylesheet" href="/algos/common.inline.css">
 		<link rel="stylesheet" href="/algos/common.async.css">
 		<script async src="/static/js/prism.js"></script>
+		<script async src="/js/common.js"></script>
+		<link rel="stylesheet" href="/js/css/common.inline.css">
 	</head>
 	<body class="pmc-u-margin-a-00">
 		<div style="display: none;">
